refactor(tests): put every test under `tests/unit` or `tests/integrations`
