+++ tests/unit/filters/test_command.py
@@ -19,7 +19,7 @@
 import pytest
 
 from pyrogram import filters
-from tests.filters import Client, Message
+from tests.unit.filters import Client, Message
 
 c = Client()
