+++ tests/unit/raw/test_optional_vector.py
